ai-engineer-mastery/requirements.txt
# Core LLM APIs
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0

# Agent Frameworks (Level 3+)
langgraph>=0.0.40
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# RAG & Vector Databases (Level 4+)
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Fine-tuning & ML (Level 5+)
transformers>=4.35.0
peft>=0.7.0
datasets>=2.15.0
bitsandbytes>=0.41.0  # For QLoRA
accelerate>=0.25.0

# Production & Monitoring (Level 6+)
fastapi>=0.110.0
uvicorn>=0.20.0
redis>=5.0.0
prometheus-client>=0.19.0
sentry-sdk>=1.40.0

# Utilities
tiktoken>=0.5.0  # Token counting
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0.0
rich>=13.0.0  # Pretty terminal output
typer>=0.9.0  # CLI framework
tqdm>=4.65.0  # Progress bars

# Testing & Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0

# Jupyter (for tutorials)
jupyter>=1.0.0
ipywidgets>=8.0.0
